Fuzz and tremolo effects

+++ components/gs_effects/include/gs_effect_bypass.h
@@ -2,4 +2,4 @@
 #include <stdbool.h>
 #include <stddef.h>
 
-void gs_bypass_effect(int32_t *input_buffer, size_t input_bytes_read, int32_t *output_buffer);
+void gs_bypass_effect(int16_t *input_buffer, size_t input_bytes_read, int16_t *output_buffer);
